+++ api/routes/translation_version.json
@@ -1,1 +1,1 @@
-{"version": "1.21", "hash": "0aa126206771f5f7855dc4fc376f3241"}
+{"version": "1.22", "hash": "30e0ae6b2c2f4c1539222f991e6dfb3c"}
