+++ frontend/src/scss/languageToggleContainer.scss
@@ -1,7 +1,7 @@
 .language-switcher-container {
   display: flex;
   align-items: center;
-  margin-right: 3rem;
+  margin-right: 1rem;
 }
 
 .language-button {
