+++ frontend/src/scss/media480.scss
@@ -2,9 +2,6 @@
   header{
 
     .logo-btn{
-      left: 0;
-      margin-left: 1rem;
-      width: 10rem;
     }
   }
