+++ frontend/src/scss/media768.scss
@@ -10,17 +10,28 @@
 
   header{
 
-    .logo-btn{
-      left: 0;
-      margin: 0 1rem;
-      transform: none;
-    }
   }
 
   .navbar-outer{
 
+    .navbar-box{
+
+      .btn-controller-box{
+        padding: 0.2rem 0;
+        margin-right: 1rem;
+
+        .nav-link{
+          padding: 0.45rem;
+        }
+      }
+    }
+
     .navbar{
       margin-left: 0;
+
+      .nav-link{
+        padding: 0.7rem 1.3rem;
+      }
     }
   }
